Fix for MacOS
