+++ CMakeLists.txt
@@ -2,7 +2,8 @@ cmake_minimum_required (VERSION 3.1)
 
 project(ubinder)
 
-set(CMAKE_CXX_STANDARD 14)
+list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 set(ubinder_BUILDING_WRAPPER OFF CACHE BOOL "If ON, means we build wrapper, not binding")
