+++ examples/key_value/gen_proto.sh
@@ -1,1 +1,2 @@
-protoc -I ./proto --java_out=./java/JavaKV/src/main/java/messages --js_out=import_style=commonjs,binary:./node/generated ./proto/messages.proto 
+mkdir -p ./java/JavaKV/src/main/java
+protoc -I ./proto --java_out=./java/JavaKV/src/main/java --js_out=import_style=commonjs,binary:./node/generated ./proto/messages.proto
