+++ examples/key_value/wrapped_lib/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(wrapped_key_value_storage)
 
 find_package(protobuf CONFIG REQUIRED)
-find_package(stlab CONFIG REQUIRED)
+find_package(stlab REQUIRED)
 
 set(PATH_TO_PROTO ${CMAKE_CURRENT_SOURCE_DIR}/../proto)
 file(GLOB ProtoFiles "${PATH_TO_PROTO}/*.proto")
@@ -13,6 +13,13 @@ protobuf_generate(
     TARGET wrapped_lib
     APPEND_PATH ${PATH_TO_PROTO}
     PROTOS ${ProtoFiles})
+
+protobuf_generate(TARGET ${executable_name})
+#protobuf_generate_cpp(
+#        wrapped_lib.cpp
+#        wrapped_lib.h
+#        ${ProtoFiles}
+#)
 
 target_link_libraries(wrapped_lib PUBLIC key_value_storage)
 target_link_libraries(wrapped_lib PRIVATE
