+++ include/ubinder/function_types.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <stdint.h>
-
+#include <stdio.h>
 #ifdef __cplusplus
 extern "C" {
 #endif
